+++ .env.local
@@ -1,4 +1,4 @@
-NEXT_POSTGRESQL_URI=postgres://postgres:xtKyTbWXCNVQsj3Q@db.juvxtsyifhdynkicgqhn.supabase.co:5432/postgres
+POSTGRESQL_URI=postgres://postgres:xtKyTbWXCNVQsj3Q@db.juvxtsyifhdynkicgqhn.supabase.co:5432/postgres
 NEXT_PUBLIC_SUPABASE_URL=https://juvxtsyifhdynkicgqhn.supabase.co
 NEXT_PUBLIC_SUPABASE_KEY=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJzdXBhYmFzZSIsInJlZiI6Imp1dnh0c3lpZmhkeW5raWNncWhuIiwicm9sZSI6ImFub24iLCJpYXQiOjE2OTA0MDI0NzYsImV4cCI6MjAwNTk3ODQ3Nn0.2_QgIzW420tzN9VNqPDaFeQdmia0eYsp3_IgDXpx3dQ
 NEXT_SUPABASE_STORAGE_URL=https://juvxtsyifhdynkicgqhn.supabase.co/storage/v1
