+++ app/api/content/[contentId]/route.ts
@@ -1,4 +1,4 @@
-import { db } from "@/libs/drizzle/db";
+import { db } from '@/db/drizzle'
 
 export async function GET(req: Request) {
   const { searchParams } = new URL(req.url);
