+++ db/schema-out.ts/meta/0000_snapshot.json
@@ -1,7 +1,7 @@
 {
   "version": "5",
   "dialect": "pg",
-  "id": "2db3720c-026a-4cbf-b9b1-d43578d07c16",
+  "id": "d695415e-e031-4a17-a519-f979a6d2b898",
   "prevId": "00000000-0000-0000-0000-000000000000",
   "tables": {
     "addresses": {
